LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-named-scroll-progress-timeline.tentative-expected.txt


PASS scroll-timeline-name is referenceable in animation-timeline on the declaring element itself
PASS scroll-timeline-name is referenceable in animation-timeline on that element's descendants
PASS scroll-timeline-name is not referenceable in animation-timeline on that element's siblings
PASS scroll-timeline-name on an element which is not a scroll-container
PASS Change in scroll-timeline-name to match animation timeline updates animation.
PASS Change in scroll-timeline-name to no longer match animation timeline updates animation.
PASS Timeline lookup updates candidate when closer match available.
PASS Timeline lookup updates candidate when match becomes available.
PASS scroll-timeline-axis is block
PASS scroll-timeline-axis is inline
PASS scroll-timeline-axis is x
PASS scroll-timeline-axis is y
PASS scroll-timeline-axis is mutated
